requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
discord.py==2.3.2
exceptiongroup==1.2.1
ffmpeg==1.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.9.4
youtube-dl==2021.12.17
youtube-search-python==1.6.6
